==> 02_code/00_preprocess.Rmd <==
---
title: "preprocess"
output: html_document
---

```{r setup, include=FALSE}
library(tidyverse)
library(mlr)
library(palabmod)
```

Set up the data paths and load the positive and negative data that are post stratification. The training data are loaded first; the positive training data represents 80% of the complete positive cohort and the negative cohort 
```{r}
data_path <- "Z:/"

pos <- readr::read_csv(file.path(data_path, "pos_cohort_20170315/poscoh_80perc_consumer_v1.csv"))
neg <- readr::read_csv(file.path(data_path, "neg_cohort_20170328/negcoh_perc80_20170320.csv"))
var_config <- readr::read_csv(file.path(data_path, "pos_cohort_20170315/Gilead_var_config2.csv"))

```
Strip ids and matching variable. Subset cohorts to relevant modelling variables. 
```{r}
pos_id <- utils_get_ids(pos, var_config)
neg_id <- utils_get_ids(neg, var_config)
pos <- utils_get_variables(pos,var_config)
neg <- utils_get_variables(neg,var_config)
extra_cols <- setdiff(colnames(neg),colnames(pos))
```

```{r}
data <- rbind(pos,neg)

```
