==> 02_code/00_setup_packages_for_palabmod.R <==
install.packages("mlr")
install.packages("BBmisc")
install.packages("ROCR")
install.packages("plotmo")
install.packages("gridExtra")
install.packages("parallelMap")
install.packages("rattle")
install.packages("rpart.plot")
install.packages("fastcluster")
install.packages("yakmoR")
install.packages("fields")
install.packages("hash")
