==> 02_code/07_xgboost_gridsearch_freq_counts.Rmd <==
---
title: "LR"
output: html_document
author: "Orla Doyle - PA Team, London"
---
Set up packages. 
```{r}
library(tidyverse)
library(mlr)
library(parallel)
library(parallelMap)
library(ggplot2)
library(palabmod)
```

Set up paths and load the data and var_config file. 
```{r}
path = "F:/orla/HCV_manuscript/"
source(file.path(path, "02_code/utils.R"))
data_tr <- readRDS(file.path(path, "01_data/gilead_train_pos_neg.rds"))
var_config <- read_csv(file.path(path, "01_data/meta_data/hcv_varconfig_counts_freq.csv"))
```

Set up matching, pull ids and variables that should go forward for modelling. 
```{r}
data_tr$test_patient_id[data_tr$label==1] <- data_tr$Patient_Id[data_tr$label==1]
matches <- as.factor(data_tr$test_patient_id)
ids <- utils_get_ids(data_tr, var_config)
data_tr <- utils_get_variables(data_tr, var_config, categorical = TRUE)
data_tr$label <- (data_tr$label)
data_tr$PAT_GENDER_CD = as.numeric(data_tr$PAT_GENDER_CD)-1
```

#Missing data
Set missing values to zero. 
```{r}
# vars = c("COUNT_OF_TX_FOR_MISDIAGNOSIS_COMORBIDITIES_SYMPTOMS_RISK_FACTORS",
#  "CLAIM_COUNT_OF_TX_FOR_MISDIAGNOSIS_COMORBIDITIES_SYMPTOMS_RISK_FACTORS",
#  "COUNT_OF_TX_PROC_FOR_MISDIAGNOSIS_COMORBIDITIES_SYMPTOMS_RISK_FACTORS",
#  "CLAIM_COUNT_OF_TX_PROC_FOR_MISDIAGNOSIS_COMORBIDITIES_SYMPTOMS_RISK_FACTORS")

# data_tmp = data_tr[, vars]
# data_tmp[is.na(data_tmp)] = 0
# data_tr[, vars] = data_tmp

data_tr[is.na(data_tr)] <- 0

```

Define MLR compatible dataset. 
```{r}
target = "label"
dataset_tr <- makeClassifTask(id="hcv", data=data_tr, target=target, 
                             positive=1, blocking=matches)
```

Define the learner 
```{r}
xgb <- makeLearner("classif.xgboost", predict.type="prob")
xgb$par.vals = list(
  nrounds = 50,
  verbose = T,
  objective = "binary:logistic"
)
```

Define the hyperparameters and metrics

```{r}
random_seed <- 123
recall_thrs <- 50
random_search_iter <- 20L
cpu_num = 20

ps = makeParamSet(
  makeNumericParam("eta", lower=0.01, upper=0.3),
  makeIntegerParam("max_depth", lower=2, upper=6),
  makeIntegerParam("min_child_weight", lower=1, upper=5),
  makeNumericParam("colsample_bytree", lower=.5, upper=1),
  makeNumericParam("subsample", lower=.5, upper=1)
)

ctrl <- makeTuneControlRandom(maxit=random_search_iter, tune.threshold=F)

# Define performane metrics - use at least 2, otherwise get_results won't work
pr10 <- perf_make_pr_measure(recall_thrs, "pr10")
m2 <- auc
m3 <- setAggregation(pr10, test.sd)
m4 <- setAggregation(auc, test.sd)
# It's always the first in the list that's used to rank hyperparams in tuning
m_all <- list(pr10, m2, m3, m4)
```

```{r}
inner <- makeResampleDesc("CV", iters=3, stratify=T)
outer <- makeResampleDesc("CV", iters=2, stratify=T, predict = "both")

outer$stratify <- FALSE
inner$stratify <- FALSE

# Define wrapped learner: this is mlR's way of doing nested CV on a learner
xgb_wrap <- makeTuneWrapper(xgb, resampling=inner, par.set=ps, control=ctrl,
                            show.info=F, measures=m_all)


```
Train with CV to find hyperparams

```{r}
parallelStartSocket(detectCores(), level="mlr.tuneParams")
res <- resample(xgb_wrap, dataset_tr, resampling=outer, models=T,
                extract=getTuneResult, show.info=F, measures=m_all)
parallelStop()
readr::write_rds(res, file.path(path, "03_results/xgb/all_results_inner.rds"))
```



Load first set of hold out data and set up matching variable and selection modelling variables. 
```{r}
data_ts <- readRDS(file.path(path, "01_data/gilead_holdout1_pos_neg.rds"))
data_ts$PAT_GENDER_CD = as.numeric(data_ts$PAT_GENDER_CD)-1
ts_matches <- as.factor(data_ts$test_patient_id)
#ids <- utils_get_ids(data_ts, var_config)
data_ts <- utils_get_variables(data_ts, var_config, categorical = TRUE)
data_ts$label <- as.factor(data_ts$label)
```
Make MLR data object from test data. 
```{r}
dataset_ts <- makeClassifTask(id="hcv_ts", data=data_ts, target="label",
                                 positive=1, blocking=ts_matches)
```
Apply the model to the test data. 
```{r}
pred_ts <- predict(xgb.fit, dataset_ts)
write.csv(as.data.frame(pred_ts), file.path(path,"03_results/xgboost/predictions_ts_xgb_misclasspen_HCV_counts_freq_1_50.csv"))
pr_grpl <- perf_binned_perf_curve(pred_ts, 20,"rec", "prec")
plot(pr_grpl$curve$rec_binned, pr_grpl$curve$prec)
write.csv(pr_grpl$curve, file.path(path,"03_results/xgboost/PR_xgb_misclasspen_HCV_counts_freq_1_50.csv"))
```


<!-- Load second set of hold out data and set up matching variable and selection modelling variables.  -->
<!-- ```{r} -->
<!-- data_ts <- readRDS(file.path(path, "01_data/gilead_holdout2_pos_neg.rds")) -->
<!-- data_ts$PAT_GENDER_CD = as.numeric(data_ts$PAT_GENDER_CD)-1 -->
<!-- ts_matches <- as.factor(data_ts$test_patient_id) -->
<!-- #ids <- utils_get_ids(data_ts, var_config) -->
<!-- data_ts <- utils_get_variables(data_ts, var_config, categorical = TRUE) -->
<!-- data_ts$label <- as.factor(data_ts$label) -->
<!-- ``` -->
<!-- Make MLR data object from test data.  -->
<!-- ```{r} -->
<!-- dataset_ts <- makeClassifTask(id="hcv_ts", data=data_ts, target="label", -->
<!--                                  positive=1, blocking=ts_matches) -->
<!-- ``` -->
<!-- Apply the model to the test data.  -->
<!-- ```{r} -->
<!-- pred_ts <- predict(xgb.fit, dataset_ts) -->
<!-- write.csv(as.data.frame(pred_ts), file.path(path,"03_results/xgboost/predictions_ts_xgb_misclasspen_HCV_counts_freq_1_50_2.csv")) -->
<!-- pr_grpl <- perf_binned_perf_curve(pred_ts, 20,"rec", "prec") -->
<!-- plot(pr_grpl$curve$rec_binned, pr_grpl$curve$prec) -->
<!-- write.csv(pr_grpl$curve, file.path(path,"03_results/xgboost/PR_xgb_misclasspen_HCV_counts_freq_1_50_2.csv")) -->
<!-- ``` -->

Get variable importance table
```{r}
#vi = results_vi_table(xgb.fit, dataset_ts)
vi <- xgboost::xgb.importance(feature_names=colnames(data_tr), model=xgb.fit$learner.model)
write.csv(vi, paste0(path, "03_results/xgb/VI_xgb_misclasspen_HCV_counts_counts_freq_1_50.csv"))
```
